+++ src/main/java/com/uniminuto/clinica/repository/SessionRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
  * Repositorio para la entidad Session.
  */
 @Repository
-public interface SessionRepository extends JpaRepository<Session, Integer> {
-    void deleteByUserId(Integer userId);
+public interface SessionRepository extends JpaRepository<Session, Long> {
+    void deleteByUserId(Long userId);
 }
